@@ -11,27 +11,27 @@
     "lint": "npm run eslint"
   },
   "dependencies": {
-    "@floating-ui/react-dom": "0.6.3",
-    "@floating-ui/react-dom-interactions": "0.3.1",
-    "@testing-library/jest-dom": "5.16.1",
-    "@testing-library/react": "12.1.2",
-    "@testing-library/user-event": "13.5.0",
-    "@types/jest": "27.0.3",
-    "@types/node": "17.0.4",
-    "@types/react": "17.0.38",
-    "@types/react-dom": "17.0.11",
-    "axios": "0.25.0",
+    "@floating-ui/react-dom": "0.7.0",
+    "@floating-ui/react-dom-interactions": "0.6.0",
+    "@testing-library/jest-dom": "5.16.4",
+    "@testing-library/react": "13.2.0",
+    "@testing-library/user-event": "14.2.0",
+    "@types/jest": "27.5.1",
+    "@types/node": "17.0.34",
+    "@types/react": "18.0.9",
+    "@types/react-dom": "18.0.4",
+    "axios": "0.27.2",
     "match-sorter": "6.3.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-scripts": "5.0.0",
-    "sass": "1.45.1",
-    "typescript": "4.5.4"
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
+    "react-scripts": "5.0.1",
+    "sass": "1.51.0",
+    "typescript": "4.6.4"
   },
   "devDependencies": {
-    "eslint": "8.5.0",
+    "eslint": "8.15.0",
     "eslint-config-airbnb-typescript-prettier": "5.0.0",
-    "prettier": "2.5.1"
+    "prettier": "2.6.2"
   },
   "prettier": {
     "printWidth": 100,
