@@ -1,5 +1,9 @@
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import './styles/global.scss';
 import Root from './components/Root';
 
-ReactDOM.render(<Root />, document.getElementById('root'));
+const root = document.getElementById('root');
+
+if (root) {
+  createRoot(root).render(<Root />);
+}
